Version 1.0
Run Melbourne_CBD to extract just the lats and lons of SA1 centroids in the CBD. Then run simpler_distance_calculation to get a csv file of lats and lons and polling place addresses. Then run send_to_distance_matrix to get the travel times. Not a best practice, I will work on refining this somewhat. You need to put dataframe11, dataframe16 and dataframe21 into the Data folder for this to work.